Extracted gyro as a subsystem.
