+++ src/org/usfirst/frc/team6179/robot/LogitechGamepad.java
@@ -30,6 +30,10 @@ public class LogitechGamepad implements OI {
         new JoystickButton(stick, LogitechGamepadKeyMapping.BTN_RB).whileHeld(new AimMode());
         // bind buttons to commands. //
 
+        // set parameters for drive straight test
+        SmartDashboard.putNumber("Speed", 0.8);
+        SmartDashboard.putNumber("Turning Increment", 0.5);
+
         // display commands on dashboard for easy testing. //
         // Shooter
         SmartDashboard.putData("Shoot Boulder", new ShootBoulder());
