+++ src/org/usfirst/frc/team6179/robot/subsystems/DriveTrain.java
@@ -2,13 +2,9 @@
 package org.usfirst.frc.team6179.robot.subsystems;
 
 import edu.wpi.first.wpilibj.RobotDrive;
-import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import org.usfirst.frc.team6179.robot.commands.drivetrain.ArcadeDriveWithJoystick;
-import org.usfirst.frc.team6179.robot.commands.drivetrain.UseGyro;
-import org.usfirst.frc.team6179.robot.configurations.DriveTrainConfig;
 import org.usfirst.frc.team6179.robot.mappings.RobotMap;
-import org.usfirst.frc.team6179.robot.sensors.GY521GyroAccelerometer;
 
 /**
  *
@@ -24,18 +20,11 @@ public class DriveTrain extends Subsystem {
     // Indicates whether the speed input should be squared
     public boolean squaredInput = true;
 
-    // private BMA220Accelerometer accelerometer;
-    public GY521GyroAccelerometer gy521;
-
-    public double angle;
-
     private RobotDrive drive;
 
     public DriveTrain() {
         drive = new RobotDrive(RobotMap.leftMotor, RobotMap.rightMotor);
-        gy521 = new GY521GyroAccelerometer();
 
-        (new UseGyro()).start();
 //        accelerometer = new BMA220Accelerometer();
     }
 
@@ -51,9 +40,6 @@ public class DriveTrain extends Subsystem {
         drive.arcadeDrive(0, 0);
     }
 
-    public double getAngularVelocity() {
-        return gy521.getGyroZ() / DriveTrainConfig.gyroSensitivity - DriveTrainConfig.gyroOffset;
-    }
 
     public void initDefaultCommand() {
         setDefaultCommand(new ArcadeDriveWithJoystick());
